Fix all clippy warnings across codebase
- Remove dead code (compile_sys) in labs_tests.rs
- Fix match_same_arms, unused_self, unnecessary_wraps, needless_return
- Fix unwrap/expect usage, collapsible_match, unnecessary_cast
- Fix manual_div_ceil, needless_late_init, wildcard_imports
- Add module-level allows for JVM codegen files (expect_used, etc.)
- All 827 tests pass, clippy clean

+++ src/codegen/nasm/memory.rs
@@ -12,7 +12,7 @@ impl AsmGenerator {
         let value = &inst.operands[2];
 
         let off = match offset {
-            IrOperand::Constant(Constant::Int(v)) => *v as i64,
+            IrOperand::Constant(Constant::Int(v)) => *v,
             _ => 0,
         };
         let (base_name, _base_type) = match base {
